requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
email-validator==2.2.0
openpyxl==3.1.2
reportlab>=4.0.0
Flask-Mail>=0.9.1
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
